Removes logic from news routes, pass it to another file
